Platform to enable/disable printf
