+++ drivers/embedded/tensil/platform.h
@@ -10,6 +10,7 @@
 
 #define TENSIL_PLATFORM_ENABLE_XTIME
 #define TENSIL_PLATFORM_ENABLE_FILE_SYSTEM
+#define TENSIL_PLATFORM_ENABLE_PRINTF
 
 #define TENSIL_PLATFORM_INSTRUCTION_AXI_DMA_DEVICE_ID XPAR_AXIDMA_0_DEVICE_ID
 #define TENSIL_PLATFORM_SAMPLE_AXI_DMA_DEVICE_ID XPAR_AXIDMA_1_DEVICE_ID
@@ -32,6 +33,7 @@
 
 #define TENSIL_PLATFORM_ENABLE_XTIME
 // #define TENSIL_PLATFORM_ENABLE_FILE_SYSTEM
+#define TENSIL_PLATFORM_ENABLE_PRINTF
 
 // #define TENSIL_PLATFORM_INSTRUCTION_AXI_DMA_DEVICE_ID XPAR_AXIDMA_0_DEVICE_ID
 // #define TENSIL_PLATFORM_SAMPLE_AXI_DMA_DEVICE_ID XPAR_AXIDMA_1_DEVICE_ID
@@ -57,6 +59,7 @@ struct flash;
 int flash_read(struct flash *flash, uint8_t *buffer, size_t size, size_t *read_size);
 #define TENSIL_PLATFORM_FLASH_TYPE struct flash*
 #define TENSIL_PLATFORM_FLASH_READ flash_read
+#define TENSIL_PLATFORM_ENABLE_PRINTF
 
 #define TENSIL_PLATFORM_INSTRUCTION_AXI_DMA_DEVICE_ID XPAR_AXIDMA_0_DEVICE_ID
