+++ drivers/embedded/tensil/sample_buffer.h
@@ -36,11 +36,15 @@ bool sample_buffer_get_next_samples_ptr(
     const struct instruction_layout *layout, const uint8_t **ptr,
     uint32_t *program_counter, uint32_t *instruction_offset);
 
+#ifdef TENSIL_PLATFORM_ENABLE_PRINTF
+
 error_t sample_buffer_print_analysis(
     const struct sample_buffer *sample_buffer,
     const struct instruction_buffer *instruction_buffer,
     const struct instruction_layout *layout, bool print_summary,
     bool print_aggregates, bool print_listing, uint32_t program_counter_shift);
+
+#endif
 
 #ifdef TENSIL_PLATFORM_ENABLE_FILE_SYSTEM
